update db_loc param
