+++ microfisher/src/microfisher/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "0.8.2"
+__version__ = "0.8.6"
