+++ microfisher/src/microfisher/args_subcommand.py
@@ -1,3 +1,5 @@
+import os
+import sys
 from . import run_merge_reports
 from . import run_init_db
 from . import merging_algorithm
@@ -14,9 +16,21 @@ def init_db(args):
             print("==DEBUG== subcommand init_db completed.\n"
                   f"Prebuild database available at: {args.db_loc}\n")
 
+def parse_output_dir(args):
+    args.workspace = os.path.realpath(args.workspace)
+    try:
+        if args.out_dir != os.path.realpath(args.out_dir):
+            args.out_dir = os.path.join(os.path.realpath(args.workspace), args.out_dir)
+        os.makedirs(args.out_dir, exist_ok=True)
+    except PermissionError:
+        print(f"==Error== Permission denied to create output directory: {args.out_dir}")
+        sys.exit(-1)
+    return args
+
 
 def search_db(args):
     # print('((%s))' % args.search)
+    parse_output_dir(args)
     config = Config(args)
     centrifuge = Centrifuge(config)
     cent_kreport = CentrifugeKReport(config)
@@ -33,6 +47,7 @@ def search_db(args):
 
 def combine_reports(args):
     if not args.dry:
+        parse_output_dir(args)
         is_complete = run_merge_reports.run(args)
         if is_complete:
             print("==DEBUG== subcommand combine completed.\t"
